+++ src/UnitTestEx.Azure.Functions/Azure/Functions/FunctionTesterBase.cs
@@ -12,12 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.Net.Http;
-using System.Net.Mime;
-using System.Reflection;
-using System.Text;
 using System.Text.Json;
-using System.Threading;
 using UnitTestEx.Abstractions;
 using UnitTestEx.Hosting;
 
@@ -132,6 +127,9 @@ namespace UnitTestEx.Azure.Functions
         /// </summary>
         private IHost GetHost()
         {
+            if (_host != null)
+                return _host;
+
             lock (SyncRoot)
             {
                 if (_host != null)
@@ -186,7 +184,18 @@ namespace UnitTestEx.Azure.Functions
         }
 
         /// <inheritdoc/>
-        protected override void ResetHost() => _host = null;
+        protected override void ResetHost()
+        {
+            lock (SyncRoot)
+            {
+                if (_host is not null)
+                {
+                    _host.Dispose();
+                    _host = null;
+                    Implementor.WriteLine("The underlying UnitTestEx 'FunctionTester' Host has been reset.");
+                }
+            }
+        }
 
         /// <summary>
         /// Gets the <see cref="IServiceProvider"/> from the underlying host.
@@ -227,230 +236,6 @@ namespace UnitTestEx.Azure.Functions
         /// <typeparam name="TFunction">The Function <see cref="Type"/> that utilizes the <see cref="ServiceBusTriggerAttribute"/> to be tested.</typeparam>
         /// <returns>The <see cref="ServiceBusTriggerTester{TFunction}"/>.</returns>
         public ServiceBusTriggerTester<TFunction> ServiceBusTrigger<TFunction>() where TFunction : class => new(this, HostExecutionWrapper(() => GetHost().Services.CreateScope()));
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with no body.
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri)
-#else
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri)
-#endif
-            => CreateHttpRequest(httpMethod, requestUri, null, MediaTypeNames.Text.Plain, null);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with <paramref name="body"/> (defaults <see cref="HttpRequest.ContentType"/> to <see cref="MediaTypeNames.Text.Plain"/>).
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="body">The optional body content.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, string? body)
-#else
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri, string? body)
-#endif
-            => CreateHttpRequest(httpMethod, requestUri, body, null, null);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with <paramref name="body"/> and <paramref name="contentType"/>.
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="body">The optional body content.</param>
-        /// <param name="contentType">The content type. Defaults to <see cref="MediaTypeNames.Text.Plain"/>.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, string? body, string? contentType)
-#else
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri, string? body, string? contentType)
-#endif
-            => CreateHttpRequest(httpMethod, requestUri, body, contentType, null);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with no body.
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="requestModifier">The optional <see cref="HttpRequest"/> modifier.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, Action<HttpRequest>? requestModifier = null)
-#else
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri, Action<HttpRequest>? requestModifier = null)
-#endif
-            => CreateHttpRequest(httpMethod, requestUri, null, MediaTypeNames.Text.Plain, requestModifier);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with <paramref name="body"/> (defaults <see cref="HttpRequest.ContentType"/> to <see cref="MediaTypeNames.Text.Plain"/>).
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="body">The optional body content.</param>
-        /// <param name="requestModifier">The optional <see cref="HttpRequest"/> modifier.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, string? body, Action<HttpRequest>? requestModifier = null)
-#else
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri, string? body, Action<HttpRequest>? requestModifier = null)
-#endif
-            => CreateHttpRequest(httpMethod, requestUri, body, null, requestModifier);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with <i>optional</i> <paramref name="body"/> (defaults <see cref="HttpRequest.ContentType"/> to <see cref="MediaTypeNames.Text.Plain"/>).
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="body">The optional body content.</param>
-        /// <param name="contentType">The content type. Defaults to <see cref="MediaTypeNames.Text.Plain"/>.</param>
-        /// <param name="requestModifier">The optional <see cref="HttpRequest"/> modifier.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri = null, string? body = null, string? contentType = null, Action<HttpRequest>? requestModifier = null)
-#else
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri = null, string? body = null, string? contentType = null, Action<HttpRequest>? requestModifier = null)
-#endif
-        {
-            if (httpMethod == HttpMethod.Get && body != null)
-                LoggerProvider.CreateLogger("FunctionTesterBase").LogWarning("A payload within a GET request message has no defined semantics; sending a payload body on a GET request might cause some existing implementations to reject the request (see https://www.rfc-editor.org/rfc/rfc7231).");
-
-            var context = new DefaultHttpContext();
-
-            var uri = requestUri is null ? new Uri("http://functiontest") : new Uri(requestUri, UriKind.RelativeOrAbsolute);
-            if (!uri.IsAbsoluteUri)
-                uri = new Uri($"http://functiontest{(requestUri != null && requestUri.StartsWith('/') ? requestUri : $"/{requestUri}")}");
-
-            context.Request.Method = httpMethod?.Method ?? HttpMethod.Get.Method;
-            context.Request.Scheme = uri.Scheme;
-            context.Request.Host = new HostString(uri.Host);
-            context.Request.Path = uri.LocalPath;
-            context.Request.QueryString = new QueryString(uri.Query);
-
-            if (body is not null)
-            {
-                context.Request.Body = new MemoryStream(body == null ? [] : Encoding.UTF8.GetBytes(body));
-                context.Request.ContentType = contentType ?? MediaTypeNames.Text.Plain;
-            }
-            else
-                context.Request.ContentType = contentType;
-
-            requestModifier?.Invoke(context.Request);
-
-            if (SetUp.OnBeforeHttpRequestSendAsync != null)
-                SetUp.OnBeforeHttpRequestSendAsync(context.Request, UserName, CancellationToken.None).GetAwaiter().GetResult();
-
-            return context.Request;
-        }
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with the <paramref name="value"/> JSON serialized as <see cref="HttpRequest.ContentType"/> of <see cref="MediaTypeNames.Application.Json"/>.
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="value">The value to JSON serialize.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateJsonHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, object? value)
-#else
-        public HttpRequest CreateJsonHttpRequest(HttpMethod httpMethod, string? requestUri, object? value)
-#endif
-            => CreateJsonHttpRequest(httpMethod, requestUri, value, null);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> with the <paramref name="value"/> JSON serialized as <see cref="HttpRequest.ContentType"/> of <see cref="MediaTypeNames.Application.Json"/>.
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="value">The value to JSON serialize.</param>
-        /// <param name="requestModifier">The optional <see cref="HttpRequest"/> modifier.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateJsonHttpRequest(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, object? value, Action<HttpRequest>? requestModifier = null)
-#else
-        public HttpRequest CreateJsonHttpRequest(HttpMethod httpMethod, string? requestUri, object? value, Action<HttpRequest>? requestModifier = null)
-#endif
-        {
-            if (httpMethod == HttpMethod.Get)
-                LoggerProvider.CreateLogger("FunctionTesterBase").LogWarning("A payload within a GET request message has no defined semantics; sending a payload body on a GET request might cause some existing implementations to reject the request (see https://www.rfc-editor.org/rfc/rfc7231).");
-
-            var hr = CreateHttpRequest(httpMethod, requestUri, requestModifier: requestModifier);
-            hr.Body = new MemoryStream(new BinaryData(JsonSerializer.Serialize(value)).ToArray());
-            hr.ContentType = MediaTypeNames.Application.Json;
-            return hr;
-        }
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> using the JSON formatted embedded resource as the content (<see cref="MediaTypeNames.Application.Json"/>).
-        /// </summary>
-        /// <typeparam name="TAssembly">The <see cref="Type"/> to infer <see cref="Type.Assembly"/> for the embedded resources.</typeparam>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="resourceName">The embedded resource name (matches to the end of the fully qualifed resource name).</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateJsonHttpRequestFromResource<TAssembly>(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, string resourceName)
-#else
-        public HttpRequest CreateJsonHttpRequestFromResource<TAssembly>(HttpMethod httpMethod, string? requestUri, string resourceName)
-#endif
-            => CreateJsonHttpRequestFromResource<TAssembly>(httpMethod, requestUri, resourceName, null);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> using the JSON formatted embedded resource as the content (<see cref="MediaTypeNames.Application.Json"/>).
-        /// </summary>
-        /// <typeparam name="TAssembly">The <see cref="Type"/> to infer <see cref="Type.Assembly"/> for the embedded resources.</typeparam>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="resourceName">The embedded resource name (matches to the end of the fully qualifed resource name).</param>
-        /// <param name="requestModifier">The optional <see cref="HttpRequest"/> modifier.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateJsonHttpRequestFromResource<TAssembly>(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, string resourceName, Action<HttpRequest>? requestModifier = null)
-#else
-        public HttpRequest CreateJsonHttpRequestFromResource<TAssembly>(HttpMethod httpMethod, string? requestUri, string resourceName, Action<HttpRequest>? requestModifier = null)
-#endif
-            => CreateJsonHttpRequestFromResource(httpMethod, requestUri, resourceName, typeof(TAssembly).Assembly, requestModifier);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> using the JSON formatted embedded resource as the content (<see cref="MediaTypeNames.Application.Json"/>).
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="resourceName">The embedded resource name (matches to the end of the fully qualifed resource name).</param>
-        /// <param name="assembly">The <see cref="Assembly"/> that contains the embedded resource; defaults to <see cref="Assembly.GetEntryAssembly()"/>.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateJsonHttpRequestFromResource(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, string resourceName, Assembly assembly)
-#else
-        public HttpRequest CreateJsonHttpRequestFromResource(HttpMethod httpMethod, string? requestUri, string resourceName, Assembly assembly)
-#endif
-            => CreateJsonHttpRequestFromResource(httpMethod, requestUri, resourceName, assembly, null);
-
-        /// <summary>
-        /// Creates a new <see cref="HttpRequest"/> using the JSON formatted embedded resource as the content (<see cref="MediaTypeNames.Application.Json"/>).
-        /// </summary>
-        /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
-        /// <param name="requestUri">The requuest uri.</param>
-        /// <param name="resourceName">The embedded resource name (matches to the end of the fully qualifed resource name).</param>
-        /// <param name="assembly">The <see cref="Assembly"/> that contains the embedded resource; defaults to <see cref="Assembly.GetEntryAssembly()"/>.</param>
-        /// <param name="requestModifier">The optional <see cref="HttpRequest"/> modifier.</param>
-        /// <returns>The <see cref="HttpRequest"/>.</returns>
-#if NET7_0_OR_GREATER
-        public HttpRequest CreateJsonHttpRequestFromResource(HttpMethod httpMethod, [StringSyntax(StringSyntaxAttribute.Uri)] string? requestUri, string resourceName, Assembly assembly, Action<HttpRequest>? requestModifier = null)
-#else
-        public HttpRequest CreateJsonHttpRequestFromResource(HttpMethod httpMethod, string? requestUri, string resourceName, Assembly assembly, Action<HttpRequest>? requestModifier = null)
-#endif
-        {
-            if (httpMethod == HttpMethod.Get)
-                LoggerProvider.CreateLogger("FunctionTesterBase").LogWarning("A payload within a GET request message has no defined semantics; sending a payload body on a GET request might cause some existing implementations to reject the request (see https://www.rfc-editor.org/rfc/rfc7231).");
-
-            var hr = CreateHttpRequest(httpMethod, requestUri, requestModifier: requestModifier);
-            hr.Body = new MemoryStream(Encoding.UTF8.GetBytes(Resource.GetJson(resourceName, assembly ?? Assembly.GetCallingAssembly())));
-            hr.ContentType = MediaTypeNames.Application.Json;
-            return hr;
-        }
 
         /// <summary>
         /// Releases all resources.